CPUTwo port of XV6, starts to boot, but fails on (not yet implemented in emulator) readwrite I/O

+++ kernel/types.h
@@ -7,4 +7,4 @@ typedef unsigned short uint16;
 typedef unsigned int  uint32;
 typedef unsigned long uint64;
 
-typedef uint64 pde_t;
+typedef uint32 pde_t;
